Make total results a long
[#165930666]

+++ src/main/java/com/boclips/events/types/video/VideosSearched.java
@@ -26,5 +26,5 @@ public class VideosSearched extends Event implements UserEvent {
     private Integer pageSize;
 
     @NonNull
-    private Integer totalResults;
+    private Long totalResults;
 }
